Fix warning

+++ 3dEngine/src/scene/renderer3d/ambientocclusion/AmbientOcclusionManager.cpp
@@ -1,8 +1,6 @@
 #include <map>
 #include <string>
-#include <cmath>
 #include <locale>
-#include <limits>
 #include <random>
 
 #include "AmbientOcclusionManager.h"
@@ -201,7 +199,7 @@ namespace urchin
                     randomFloats(generator));
             sample = sample.normalize();
 
-            float scale = static_cast<float>(i) / kernelSamples;
+            float scale = static_cast<float>(i) / static_cast<float>(kernelSamples);
             scale = MathAlgorithm::lerp<float>(0.1f, 1.0f, scale * scale); //use square function to bring most of sample closer to center
             sample *= scale;
             ssaoKernel.push_back(sample);
